Refactor import script and enhance project structure
Refactored `import_csv.py` to improve reusability and eliminated hardcoded paths. Updated SQL schema, added new genres/movies data, and modularized SQL scripts for functions/procedures. Enhanced documentation and restructured project files for clarity and scalability.

+++ M164_GIBZ_Projektareite/sql/create_tables.sql
@@ -18,7 +18,7 @@ CREATE TABLE Users (
 );
 
 CREATE TABLE Ratings (
-    RatingID INT PRIMARY KEY,
+    RatingID BIGINT PRIMARY KEY,
     MovieID INT,
     UserID INT,
     Score DECIMAL(2,1),
@@ -28,7 +28,7 @@ CREATE TABLE Ratings (
 );
 
 CREATE TABLE Reviews (
-    ReviewID INT PRIMARY KEY,
+    ReviewID BIGINT PRIMARY KEY,
     MovieID INT,
     UserID INT,
     Text TEXT,
